yuhao_requirements.txt

fast-slic==0.4.0
huggingface-hub
imageio==2.31.1
importlib-metadata
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy 
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
Pillow==9.4.0
protobuf==3.20.3
scikit-image==0.21.0
scipy==1.11.1
torch
torchaudio==2.0.0
torchvision
tornado==6.2
tqdm

